chore(conformance): extract runner tests below §19 cap
`conformance/src/runner.rs` was 2485 lines. Extract the 589-line inline
`mod tests` block into `runner_tests.rs` via `#[path]`, leaving
`runner.rs` at 1896 lines.
